http://luc.lino-framework.org/blog/2016/0816.html

+++ atelier/invlib.py
@@ -78,6 +78,7 @@ def get_current_date(today=None):
     """
 
     if today is None:
+        # return datetime.datetime.utcnow()
         return datetime.date.today()
     return i2d(today)
 
@@ -92,14 +93,15 @@ def must_exist(p):
         raise Exception("No such file: %s" % p.absolute())
 
 
-def rmtree_after_confirm(p):
+def rmtree_after_confirm(p, batch=False):
     if not p.exists():
         return
-    if confirm("OK to remove %s and everything under it?" % p.absolute()):
+    if batch or confirm(
+            "OK to remove %s and everything under it?" % p.absolute()):
         p.rmtree()
 
 
-def cleanup_pyc(p):
+def cleanup_pyc(p, batch=False):
     """Thanks to oddthinking on http://stackoverflow.com/questions/2528283
     """
     for root, dirs, files in os.walk(p):
@@ -108,37 +110,38 @@ def cleanup_pyc(p):
         excess_pyc_files = [pyc_filename for pyc_filename in pyc_files if pyc_filename[:-1] not in py_files]
         for excess_pyc_file in excess_pyc_files:
             full_path = os.path.join(root, excess_pyc_file)
-            must_confirm("Remove excess file %s:" % full_path)
-            os.remove(full_path)
+            if batch or confirm("Remove excess file %s:" % full_path):
+                os.remove(full_path)
 
 
-def sphinx_clean(ctx):
+def sphinx_clean(ctx, batch=False):
     """Delete all generated Sphinx files.
 
     """
     for docs_dir in get_doc_trees(ctx):
-        rmtree_after_confirm(docs_dir.child(ctx.build_dir_name))
+        rmtree_after_confirm(docs_dir.child(ctx.build_dir_name), batch)
 
 
-def py_clean(ctx):
+def py_clean(ctx, batch=False):
     """Delete dangling `.pyc` files.
 
     """
     if atelier.current_project.module is not None:
         p = Path(atelier.current_project.module.__file__).parent
-        cleanup_pyc(p)
+        cleanup_pyc(p, batch)
     p = ctx.root_dir.child('tests')
     if p.exists():
-        cleanup_pyc(p)
+        cleanup_pyc(p, batch)
 
     files = []
     for pat in ctx.cleanable_files:
         for p in glob.glob(os.path.join(ctx.root_dir, pat)):
             files.append(p)
     if len(files):
-        must_confirm("Remove {0} cleanable files".format(len(files)))
-        for p in files:
-            os.remove(p)
+        if batch or confirm(
+                "Remove {0} cleanable files".format(len(files))):
+            for p in files:
+                os.remove(p)
 
 
 def run_in_demo_projects(ctx, admin_cmd, *more):
@@ -253,6 +256,7 @@ def initdb_demo(ctx):
 @task(name='test')
 def run_tests(ctx):
     """Run the test suite of this project."""
+    # assert os.environ['COVERAGE_PROCESS_START']
     if not ctx.root_dir.child('setup.py').exists():
         return
     ctx.run('python setup.py -q test', pty=True)
@@ -317,10 +321,11 @@ def build_docs(ctx, *cmdline_args):
 
 
 @task(name='clean')
-def clean(ctx, *cmdline_args):
+def clean(ctx, batch=False):
+# def clean(ctx, *cmdline_args):
     """Remove temporary and generated files."""
-    sphinx_clean(ctx)
-    py_clean(ctx)
+    sphinx_clean(ctx, batch)
+    py_clean(ctx, batch)
     # clean_demo_caches()
 
 
@@ -371,44 +376,21 @@ def run_tests_coverage(ctx, html=True, html_cov_dir='htmlcov'):
     (overwriting any files without confirmation).
 
     """
-    if True:
-        SETUP_INFO = get_setup_info(ctx.root_dir)
-        if 'test_suite' not in SETUP_INFO:
-            raise Exception("No `test_suite` in your `setup.py`.")
-        test_suite = ctx.root_dir.child(SETUP_INFO['test_suite'])
-    else:
-        test_suite = ctx.root_dir
-
     covfile = ctx.root_dir.child('.coveragerc')
     if not covfile.exists():
-        raise Exception('There is no file {0}'.format(covfile))
-    import coverage
-    # ~ clean_sys_path()
-    # print("Running tests for '%s' within coverage..." % ctx.project_name)
-    print("Running tests in '%s' within coverage..." % test_suite)
-    # ~ ctx.DOCSDIR.chdir()
+        print('No .coveragerc file in {0}'.format(ctx.project_name))
+        return
+    print("Running {0} in {1} within coverage...".format(
+        ctx.coverage_command, ctx.project_name))
     os.environ['COVERAGE_PROCESS_START'] = covfile
-    # cov = coverage.coverage()
-    # cov.start()
-    # import unittest
-    # tests = unittest.TestLoader().discover(test_suite)
-    # unittest.TextTestRunner(verbosity=1).run(tests)
-    # cov.stop()
-    # cov.save()
-    # cov.combine()
-    # cov.report()
-    # htmlcov = ctx.root_dir.child('htmlcov')
-    # if htmlcov.exists():
-    #     print("Writing html report to %s" % htmlcov)
-    #     cov.html_report(include=cov.get_data().measured_files())
-    # cov.erase()
     ctx.run('coverage erase', pty=True)
-    ctx.run('coverage run setup.py test', pty=True)
+    ctx.run('coverage run {}'.format(ctx.coverage_command), pty=True)
     ctx.run('coverage combine', pty=True)
     ctx.run('coverage report', pty=True)
     if html:
         print("Writing html report to %s" % html_cov_dir)
-        ctx.run('coverage html -d {html_cov_dir} && open {html_cov_dir}/index.html'.format(html_cov_dir=html_cov_dir), pty=True)
+        ctx.run('coverage html -d {0} && open {0}/index.html'.format(
+            html_cov_dir), pty=True)
         print('html report is ready.')
     ctx.run('coverage erase', pty=True)
 
@@ -696,47 +678,9 @@ def update_catalog_code(ctx):
             ctx.run(cmd, pty=True)
 
 
-@task(name='ls')
-def list_projects(ctx, *cmdline_args):
-    """List your projects."""
-    from atelier.projects import load_projects
-    # headers = (
-    #     # ~ '#','Location',
-    #     'Project',
-    #     # 'Old version',
-    #     'Version')
-
-    headers = (
-        'Project',
-        'URL',
-        'Version',
-        'doctrees')
-
-    def cells(self):
-        self.load_fabfile()
-        yield self.nickname
-        yield self.SETUP_INFO.get('url', None)
-        yield self.SETUP_INFO.get('version', '')
-        yield ', '.join(self.doc_trees)
-
-    def old_cells(self):
-        self.load_fabfile()
-        # print 20140116, self.module
-        desc = "%s -- " % self.nickname
-        desc += "(doc_trees : %s)\n" % ', '.join(self.doc_trees)
-        url = self.SETUP_INFO.get('url', None)
-        version = self.SETUP_INFO.get('version', '')
-        if url:
-            desc += "`%s <%s>`__ -- %s" % (
-                self.name, url,
-                self.SETUP_INFO['description'])
-        return (
-            '\n'.join(textwrap.wrap(desc, 60)),
-            # self.dist.version,
-            version)
-
-    print(rstgen.table(headers, [
-        list(cells(p)) for p in load_projects()]))
+# @task(name='ls')
+# def list_projects(ctx, *cmdline_args):
+#     """List your projects."""
 
 
 @task(name='ct')
@@ -748,14 +692,20 @@ def commited_today(ctx, today=None):
     today = get_current_date(today)
     rows = []
 
-    def load(self):
+    def load(prj):
 
-        self.load_fabfile()
+        # prj.load_fabfile()
+        prj.load_tasks()
 
-        if ctx.revision_control_system != 'git':
+        # tsk, cfg = prj.ns.task_with_config('ci')
+        cfg = prj.ns.configuration()
+
+        if cfg['revision_control_system'] != 'git':
+        # if cfg.revision_control_system != 'git':
+            # print("20160816 {}".format(cfg))
             return
     
-        repo = Repo(ctx.root_dir)
+        repo = Repo(cfg['root_dir'])
 
         kw = dict()
         ONEDAY = timedelta(days=1)
@@ -765,6 +715,7 @@ def commited_today(ctx, today=None):
                   before=tomorrow.strftime("%Y-%m-%d"))
         it = list(repo.iter_commits(**kw))
         if len(it) == 0:
+            # print("20160816 no commits in {}".format(prj.nickname))
             return
 
         def fmtcommit(c):
@@ -773,17 +724,21 @@ def commited_today(ctx, today=None):
             if url.startswith("git@github.com"):
                 url = "https://github.com/" + url[15:-4] \
                       + "/commit/" + c.hexsha
+            elif url.startswith("git+ssh://git@github.com"):
+                url = "https://github.com/" + url[25:-4] \
+                      + "/commit/" + c.hexsha
             
             s = "`{0} <{1}>`__".format(c.hexsha[-7:], url)
             if c.message and not c.message.startswith("http://"):
                 s += " " + c.message
             return s
             
-        url = self.SETUP_INFO.get('url', "oops")
-        desc = "`%s <%s>`__" % (self.name, url)
+        url = prj.SETUP_INFO.get('url', "oops")
+        desc = "`%s <%s>`__" % (prj.name, url)
 
         for c in it:
-            ts = time.strftime("%H:%M", time.gmtime(c.committed_date))
+            # ts = time.strftime("%H:%M", time.gmtime(c.committed_date))
+            ts = time.strftime("%Y-%m-%d %H:%M", time.localtime(c.committed_date))
             rows.append([ts, desc, fmtcommit(c)])
 
     for p in load_projects():
